# Avoid using `len(x)` to check if a list is empty