I change the code to make the running time faster.
Travis need for the code check now 49.242s to 49.900s. The other check need 61s to 69s.
I hope that it was not an (travis)exception, it can anyone confirm?