@LinuxDonald 2016-08-12
When travis I'd faster I'd than the game maybe a bit faster too? MasterofJOKers is now 2 weeks not here (vacation)