@LinuxDonald 2016-08-12
I found the len is an python fuction https://docs.python.org/2/library/functions.html#len is len here needed or can it be really removed?