@MarkusHackspacher 2016-08-12
https://www.quantifiedcode.com/app/issue_class/3LKECaRv
or https://doc.odoo.com/contribute/15_guidelines/coding_guidelines_python/#collections-are-booleans-too