@LinuxDonald 2016-08-12
thx for the info i will test the change later. If there are no problems i will merge that :)

Thx for your work :+1: